// surveyadmin/src/core/models/user.ts
export type UserDocument = VulcanDocument & {
  isAdmin?: boolean;
  isVerified?: boolean;
  groups: Array<string>;
} & (
    | {
        /**
         * Legacy password based auth
         */
        authMode: undefined | "password";
        emailHash: string;
      }
    | { authMode: "anonymous"; emailHash?: undefined }
    | {
        authMode: "passwordless";
        emailHash: string;
      }
  );

export type UserType = UserDocument; //UserWithEmailDocument | AnonymousUserDocument;